Tip Calculator
Whenever I go out to eat and I’m having a hard time calculating the right tip, I’ll fumble with my phone and search for a “tip calculator” on Google. I couldn’t tell you the name or the URL of the one I usually end up using, but it’s a simple JavaScript app.